Big update: Expand main classes and create new WorldListener class
Delegate World drawing functionality to Renderer in RenderSim().

Create WorldListener class, that captures events happening in the World, and then the information from it is rendered by Renderer on the Log window (also in RenderSim()).

Implement more functionality in Organism and Animal classes, encapsulate fields and methods.

Reformat most of the names.